requirements.txt
requests>=2.31.0
flask>=3.1.0
python-dotenv>=1.0.1
qcodes>=0.50.1
prometheus_client>=0.21.1
werkzeug>=3.1.3
waitress>=3.0.2